Added one more filter to the inclure inc-gis_map: {limit=x}
This new filter is used to change the number of markers that will be showed. For example, if you put: {limit=10}, 10 markers will appear on the map.

Example of the inclure with the filter limit:
<INCLURE{fond=inc-gis_map}{id_rubrique=1}{id_mot}{id_auteur}{recherche}{id_article}{zoom}{latit}{lonxit}{type=carte}{control=large}{limit=3}>

We have also solved a problem with the markes, there was a visualization problem when the zoom level was low.



git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@17694 ac52e18a-acf5-0310-9fe8-c4428f23b10a